pyproject.toml
[project]
name = "active-recall"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.6.0",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langgraph-checkpoint-redis>=0.1.2",
    "load-dotenv>=0.1.0",
    "psycopg[binary,pool]>=3.2.11",
]
